+++ apps/backoffice-v2/src/pages/Root/Root.page.tsx
@@ -1,7 +1,12 @@
-import { FunctionComponent, lazy } from 'react';
+import React, { FunctionComponent, lazy, useState } from 'react';
 import { Outlet } from 'react-router-dom';
 import { Providers } from '../../common/components/templates/Providers/Providers';
 import { ServerDownLayout } from './ServerDown.layout';
+import { useCustomerQuery } from '@/domains/customer/hook/queries/useCustomerQuery/useCustomerQuery';
+import { FullScreenLoader } from '@/common/components/molecules/FullScreenLoader/FullScreenLoader';
+import Chatbot from '@/domains/chat/chatbot-opengpt';
+import { ctw } from '@/common/utils/ctw/ctw';
+import { RenderChildrenInIFrame } from '@/common/components/organisms/RenderChildrenInIFrame/RenderChildrenInIFrame';
 
 const ReactQueryDevtools = lazy(() =>
   process.env.NODE_ENV !== 'production'
@@ -11,12 +16,40 @@ const ReactQueryDevtools = lazy(() =>
     : Promise.resolve({ default: () => null }),
 );
 
+const ChatbotLayout: FunctionComponent = () => {
+  const { data: customer, isLoading: isLoadingCustomer } = useCustomerQuery();
+  const [isWebchatOpen, setIsWebchatOpen] = useState(false);
+  const toggleIsWebchatOpen = () => {
+    setIsWebchatOpen(prevState => !prevState);
+  };
+
+  if (isLoadingCustomer) {
+    return <FullScreenLoader />;
+  }
+
+  if (!customer?.config?.isChatbotEnabled) {
+    return null;
+  }
+
+  return (
+    <RenderChildrenInIFrame
+      className={ctw('fixed bottom-right-0', {
+        'h-[700px] w-[400px]': isWebchatOpen,
+        'd-[80px]': !isWebchatOpen,
+      })}
+    >
+      <Chatbot isWebchatOpen={isWebchatOpen} toggleIsWebchatOpen={toggleIsWebchatOpen} />
+    </RenderChildrenInIFrame>
+  );
+};
+
 export const Root: FunctionComponent = () => {
   return (
     <Providers>
       <ServerDownLayout>
         <Outlet />
       </ServerDownLayout>
+      <ChatbotLayout />
       {/*<Suspense>*/}
       {/*  <ReactQueryDevtools  />*/}
       {/*</Suspense>*/}